Refactor sample maker code to have more files with single class in them.
